# packages/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Pathfinder API
  description: VEuPathDB Strategy Builder Chatbot API
  version: 1.0.0
paths:
...
  /api/v1/strategies:
    get:
      tags:
      - strategies
      summary: List Strategies
      description: List user's saved strategies.
      operationId: list_strategies_api_v1_strategies_get
      security:
      - APIKeyCookie: []
      parameters:
      - name: siteId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategySummaryResponse'
                title: Response List Strategies Api V1 Strategies Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - strategies
      summary: Create Strategy
      description: Create a new strategy.
      operationId: create_strategy_api_v1_strategies_post
      security:
      - APIKeyCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/veupath_chatbot__transport__http__schemas__strategies__CreateStrategyRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/strategies/open:
    post:
      tags:
      - strategies
      summary: Open Strategy
      description: Open a strategy by local or WDK strategy.
      operationId: open_strategy_api_v1_strategies_open_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenStrategyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenStrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
  /api/v1/strategies/wdk:
    get:
      tags:
      - strategies
      summary: List Wdk Strategies
      description: List strategies from VEuPathDB WDK.
      operationId: list_wdk_strategies_api_v1_strategies_wdk_get
      parameters:
      - name: siteId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WdkStrategySummaryResponse'
                title: Response List Wdk Strategies Api V1 Strategies Wdk Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/sync-wdk:
    post:
      tags:
      - strategies
      summary: Sync All Wdk Strategies
      description: 'Batch-sync all WDK strategies into the local DB and return the full list.


        For each non-internal WDK strategy, fetches the full payload and upserts

        a local copy. Returns the complete list of local strategies for this site

        (including non-WDK drafts).'
      operationId: sync_all_wdk_strategies_api_v1_strategies_sync_wdk_post
      security:
      - APIKeyCookie: []
      parameters:
      - name: siteId
        in: query
        required: true
        schema:
          type: string
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategySummaryResponse'
                title: Response Sync All Wdk Strategies Api V1 Strategies Sync Wdk Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/wdk/{wdkStrategyId}:
    delete:
      tags:
      - strategies
      summary: Delete Wdk Strategy
      description: Delete a strategy from VEuPathDB WDK.
      operationId: delete_wdk_strategy_api_v1_strategies_wdk__wdkStrategyId__delete
      parameters:
      - name: wdkStrategyId
        in: path
        required: true
        schema:
          type: integer
          title: Wdkstrategyid
      - name: siteId
        in: query
        required: true
        schema:
          type: string
          title: Siteid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/wdk/{wdkStrategyId}/import:
    post:
      tags:
      - strategies
      summary: Import Wdk Strategy
      description: 'Import a WDK strategy as a local snapshot.


        Upserts: if a local row already exists for this WDK strategy ID

        (e.g. from a previous sync or double-click), it is updated rather

        than creating a duplicate.'
      operationId: import_wdk_strategy_api_v1_strategies_wdk__wdkStrategyId__import_post
      security:
      - APIKeyCookie: []
      parameters:
      - name: wdkStrategyId
        in: path
        required: true
        schema:
          type: integer
          title: Wdkstrategyid
      - name: siteId
        in: query
        required: true
        schema:
          type: string
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/{strategyId}/sync-wdk:
    post:
      tags:
      - strategies
      summary: Sync Strategy From Wdk
      description: Sync local strategy snapshot from VEuPathDB WDK.
      operationId: sync_strategy_from_wdk_api_v1_strategies__strategyId__sync_wdk_post
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JSONArray:
      items:
        $ref: '#/components/schemas/JSONValue'
      type: array
    JSONObject:
      additionalProperties:
        $ref: '#/components/schemas/JSONValue'
      type: object
...
    MessageResponse:
      properties:
        role:
          type: string
          title: Role
        content:
          type: string
          title: Content
        toolCalls:
          anyOf:
          - items:
              $ref: '#/components/schemas/ToolCallResponse'
            type: array
          - type: 'null'
          title: Toolcalls
        subKaniActivity:
          anyOf:
          - $ref: '#/components/schemas/SubKaniActivityResponse'
          - type: 'null'
        mode:
          anyOf:
          - type: string
          - type: 'null'
          title: Mode
        citations:
          anyOf:
          - $ref: '#/components/schemas/JSONArray'
          - type: 'null'
        planningArtifacts:
          anyOf:
          - $ref: '#/components/schemas/JSONArray'
          - type: 'null'
        reasoning:
          anyOf:
          - type: string
          - type: 'null'
          title: Reasoning
        optimizationProgress:
          anyOf:
          - $ref: '#/components/schemas/JSONObject'
          - type: 'null'
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
      - role
      - content
      - timestamp
      title: MessageResponse
      description: Chat message.
    OpenStrategyRequest:
      properties:
        strategyId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Strategyid
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
        siteId:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
      type: object
      title: OpenStrategyRequest
      description: Request to open a strategy.
    OpenStrategyResponse:
      properties:
        strategyId:
          type: string
          format: uuid
          title: Strategyid
      type: object
      required:
      - strategyId
      title: OpenStrategyResponse
      description: Open strategy response.
...
    PlanMetadata:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        siteId:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
      type: object
      title: PlanMetadata
    PlanNode-Input:
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
        displayName:
          anyOf:
          - type: string
          - type: 'null'
          title: Displayname
        filters:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepFilterSpec'
            type: array
          - type: 'null'
          title: Filters
        analyses:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepAnalysisSpec-Input'
            type: array
          - type: 'null'
          title: Analyses
        reports:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepReportSpec-Input'
            type: array
          - type: 'null'
          title: Reports
        searchName:
          type: string
          title: Searchname
        parameters:
          $ref: '#/components/schemas/JSONObject'
        primaryInput:
          anyOf:
          - $ref: '#/components/schemas/PlanNode-Input'
          - type: 'null'
        secondaryInput:
          anyOf:
          - $ref: '#/components/schemas/PlanNode-Input'
          - type: 'null'
        operator:
          anyOf:
          - type: string
          - type: 'null'
          title: Operator
        colocationParams:
          anyOf:
          - $ref: '#/components/schemas/ColocationParams'
          - type: 'null'
      additionalProperties: true
      type: object
      required:
      - searchName
      title: PlanNode
      description: 'Untyped recursive plan node (WDK-aligned).


        Kind is inferred from structure:

        - combine: primaryInput + secondaryInput

        - transform: primaryInput only

        - search: no inputs'
...
    StepAnalysisResponse:
      properties:
        analysisType:
          type: string
          title: Analysistype
        parameters:
          $ref: '#/components/schemas/JSONObject'
        customName:
          anyOf:
          - type: string
          - type: 'null'
          title: Customname
      type: object
      required:
      - analysisType
      title: StepAnalysisResponse
      description: Analysis configuration attached to a step.
...
    StepFilterResponse:
      properties:
        name:
          type: string
          title: Name
        value:
          $ref: '#/components/schemas/JSONValue'
        disabled:
          type: boolean
          title: Disabled
          default: false
      type: object
      required:
      - name
      - value
      title: StepFilterResponse
      description: Filter attached to a step.
...
    StepReportResponse:
      properties:
        reportName:
          type: string
          title: Reportname
          default: standard
        config:
          $ref: '#/components/schemas/JSONObject'
      type: object
      title: StepReportResponse
      description: Report configuration attached to a step.
...
    StepResponse:
      properties:
        id:
          type: string
          title: Id
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
        displayName:
          type: string
          title: Displayname
        searchName:
          anyOf:
          - type: string
          - type: 'null'
          title: Searchname
        recordType:
          anyOf:
          - type: string
          - type: 'null'
          title: Recordtype
        parameters:
          anyOf:
          - $ref: '#/components/schemas/JSONObject'
          - type: 'null'
        operator:
          anyOf:
          - type: string
          - type: 'null'
          title: Operator
        colocationParams:
          anyOf:
          - $ref: '#/components/schemas/JSONObject'
          - type: 'null'
        primaryInputStepId:
          anyOf:
          - type: string
          - type: 'null'
          title: Primaryinputstepid
        secondaryInputStepId:
          anyOf:
          - type: string
          - type: 'null'
          title: Secondaryinputstepid
        resultCount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resultcount
        wdkStepId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstepid
        filters:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepFilterResponse'
            type: array
          - type: 'null'
          title: Filters
        analyses:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepAnalysisResponse'
            type: array
          - type: 'null'
          title: Analyses
        reports:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepReportResponse'
            type: array
          - type: 'null'
          title: Reports
      type: object
      required:
      - id
      - displayName
      title: StepResponse
      description: Strategy step.
    StrategyPlan-Input:
      properties:
        recordType:
          type: string
          title: Recordtype
        root:
          $ref: '#/components/schemas/PlanNode-Input'
        metadata:
          anyOf:
          - $ref: '#/components/schemas/PlanMetadata'
          - type: 'null'
      additionalProperties: true
      type: object
      required:
      - recordType
      - root
      title: StrategyPlan
...
    StrategyResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        siteId:
          type: string
          title: Siteid
        recordType:
          anyOf:
          - type: string
          - type: 'null'
          title: Recordtype
        steps:
          items:
            $ref: '#/components/schemas/StepResponse'
          type: array
          title: Steps
        rootStepId:
          anyOf:
          - type: string
          - type: 'null'
          title: Rootstepid
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
        isSaved:
          type: boolean
          title: Issaved
          default: false
        messages:
          anyOf:
          - items:
              $ref: '#/components/schemas/MessageResponse'
            type: array
          - type: 'null'
          title: Messages
        thinking:
          anyOf:
          - $ref: '#/components/schemas/ThinkingResponse'
          - type: 'null'
        modelId:
          anyOf:
          - type: string
          - type: 'null'
          title: Modelid
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - name
      - siteId
      - recordType
      - steps
      - rootStepId
      - createdAt
      - updatedAt
      title: StrategyResponse
      description: Full strategy with steps.
    StrategySummaryResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        siteId:
          type: string
          title: Siteid
        recordType:
          anyOf:
          - type: string
          - type: 'null'
          title: Recordtype
        stepCount:
          type: integer
          title: Stepcount
        resultCount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resultcount
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
        isSaved:
          type: boolean
          title: Issaved
          default: false
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - name
      - siteId
      - recordType
      - stepCount
      - createdAt
      - updatedAt
      title: StrategySummaryResponse
      description: Strategy summary for list views.
...
    ThinkingResponse:
      properties:
        toolCalls:
          anyOf:
          - items:
              $ref: '#/components/schemas/ToolCallResponse'
            type: array
          - type: 'null'
          title: Toolcalls
        lastToolCalls:
          anyOf:
          - items:
              $ref: '#/components/schemas/ToolCallResponse'
            type: array
          - type: 'null'
          title: Lasttoolcalls
        subKaniCalls:
          anyOf:
          - additionalProperties:
              items:
                $ref: '#/components/schemas/ToolCallResponse'
              type: array
            type: object
          - type: 'null'
          title: Subkanicalls
        subKaniStatus:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Subkanistatus
        reasoning:
          anyOf:
          - type: string
          - type: 'null'
          title: Reasoning
        updatedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updatedat
      type: object
      title: ThinkingResponse
      description: In-progress tool call state.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WdkStrategySummaryResponse:
      properties:
        wdkStrategyId:
          type: integer
          title: Wdkstrategyid
        name:
          type: string
          title: Name
        siteId:
          type: string
          title: Siteid
        wdkUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Wdkurl
        rootStepId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rootstepid
        isSaved:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Issaved
        isInternal:
          type: boolean
          title: Isinternal
          default: false
      type: object
      required:
      - wdkStrategyId
      - name
      - siteId
      title: WdkStrategySummaryResponse
      description: WDK strategy summary for list views.
...
    veupath_chatbot__transport__http__schemas__strategies__CreateStrategyRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        siteId:
          type: string
          title: Siteid
        plan:
          $ref: '#/components/schemas/StrategyPlan-Input'
      type: object
      required:
      - name
      - siteId
      - plan
      title: CreateStrategyRequest
      description: Request to create a strategy.
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: pathfinder-auth
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.pathfinder.veupathdb.org
  description: Production